+++ src/net/sf/freecol/server/generator/SimpleMapGenerator.java
@@ -764,7 +764,7 @@ public class SimpleMapGenerator implements MapGenerator {
         for (Tile t: tile.getSurroundingTiles(1)) {
             for (Entry<GoodsType, Integer> entry : scale.entrySet()) {
                 GoodsType goodsType = entry.getKey();
-                scale.put(goodsType, entry.getValue().intValue()
+                scale.put(goodsType, entry.getValue()
                           + t.getPotentialProduction(goodsType, null));
             }
         }
@@ -772,7 +772,7 @@ public class SimpleMapGenerator implements MapGenerator {
         List<RandomChoice<UnitType>> scaledSkills = new ArrayList<>();
         for (RandomChoice<UnitType> skill : skills) {
             UnitType unitType = skill.getObject();
-            int scaleValue = scale.get(unitType.getExpertProduction()).intValue();
+            int scaleValue = scale.get(unitType.getExpertProduction());
             scaledSkills.add(new RandomChoice<>(unitType,
                     skill.getProbability() * scaleValue));
         }
